renamed ftrl_proximal to ftrl
